[z1839] Implement Creature Linking via database
Thanks to Silverice for feedback!

This system interprets the content of the table `creature_linking_template`. To trigger different actions on different events of the npcs that are linked together.
Possible event/ action combinations can be taken form the flags in CreatureLinkingMgr.h::CreatureLinkingFlags

(based on commit a81eb7f)
(based on commit 0a0540e)
(based on commit 6170e3d)